+++ packages/components/src/DataList/DataList.css.d.ts
@@ -13,6 +13,7 @@ declare const styles: {
   readonly "selected": string;
   readonly "filtering": string;
   readonly "filteringSpinner": string;
+  readonly "srOnly": string;
 };
 export = styles;