+++ packages/components/src/DataList/DataList.tsx
@@ -26,6 +26,8 @@ import { DataListAction } from "./components/DataListAction";
 import { DataListLayoutActions } from "./components/DataListLayoutActions";
 import { DataListContext, useDataListContext } from "./context/DataListContext";
 import {
+  DataListBulkActionProps,
+  DataListBulkActionsProps,
   DataListEmptyStateProps,
   DataListFiltersProps,
   DataListItemActionsProps,
@@ -41,6 +43,7 @@ import {
 } from "./DataList.utils";
 import { useLayoutMediaQueries } from "./hooks/useLayoutMediaQueries";
 import { DATA_LIST_FILTERING_SPINNER_TEST_ID } from "./DataList.const";
+import { DataListBulkActions } from "./components/DataListBulkActions";
 import { Heading } from "../Heading";
 import { Spinner } from "../Spinner";
 
@@ -66,6 +69,10 @@ export function DataList<T extends DataListObject>({
   const itemActionComponent = getCompoundComponent<
     DataListItemActionsProps<DataListObject>
   >(props.children, DataListItemActions);
+  const bulkActionsComponent = getCompoundComponent<DataListBulkActionsProps>(
+    props.children,
+    DataListBulkActions,
+  );
 
   return (
     <DataListContext.Provider
@@ -75,6 +82,7 @@ export function DataList<T extends DataListObject>({
         layoutComponents,
         emptyStateComponents,
         itemActionComponent,
+        bulkActionsComponent,
         ...props,
         selected: props.selected ?? [],
         // T !== DataListObject
@@ -215,7 +223,22 @@ DataList.Search = DataListSearch;
 DataList.ItemActions = DataListItemActions;
 
 /**
+ * Defines the group actions you could do on multiple DataList items.
+ */
+DataList.BulkActions = DataListBulkActions;
+
+/**
  * Defines the action in a DataList. This should be used inside the
  * DataListItemActions component.
  */
-DataList.Action = DataListAction;
+DataList.ItemAction = DataListAction;
+
+/**
+ * Defines the batch action in a DataList. This should be used inside the
+ * DataListBulkActions component.
+ */
+DataList.BatchAction = function DataListBatchAction(
+  props: DataListBulkActionProps,
+) {
+  return <DataListAction {...props} />;
+};