+++ packages/components/src/DataList/components/DataListBulkActions/DataListBulkActions.const.ts
@@ -1,0 +1,2 @@
+export const BULK_ACTIONS_CONTAINER_TEST_ID =
+  "ATL-DataListBulkActions-Container";